+++ app/Http/Controllers/PostsController.php
@@ -6,7 +6,9 @@ use App\Models\Posts;
 use App\Models\Categories;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 
 class PostsController extends Controller
 {
@@ -30,12 +32,18 @@ class PostsController extends Controller
 
     public function addPost()
     {
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         $all_categories = Categories::all();
         return view('add_post', ['categories' => $all_categories]);
     }
 
     public function savePost(Request $request)
     {
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         $path_img = $request->file('image')->store('uploads', 'public_img');
 
         Posts::create([
@@ -56,6 +64,9 @@ class PostsController extends Controller
 
     public function showAllPosts(User $user)
     {
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         $posts = Posts::all();
         $categories = $this->getCategories();
         return view('all_posts_admin', ['posts' => $posts, 'categories' => $categories]);
@@ -76,13 +87,18 @@ class PostsController extends Controller
 
     public function showEditPostForm(Posts $post)
     {
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         $categories = Categories::all();
         return view('edit_post', ['post' => $post, 'categories' => $categories]);
     }
 
     public function updatePost(Request $request, Posts $post)
     {
-
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         $array_post_data = [
             'title' => $request->title,
             'preview' => $request->preview,
@@ -107,11 +123,17 @@ class PostsController extends Controller
 
     public function showDeletePostForm(Posts $post)
     {
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         return view('delete_post', ['post' => $post]);
     }
 
     public function deletePost(Posts $post)
     {
+        if(!Gate::allows('access-admin')) {
+            abort(403);
+        }
         $post->delete();
         return redirect()->route('home');
     }
